Error fixing

+++ firmware/Core/Src/obc_systems/adc_vbat_monitor.c
@@ -4,11 +4,11 @@
 
 
 
-float return_VBAT_Voltage(){
+double return_VBAT_Voltage(){
 
     HAL_ADC_Start(&hadc1);
     HAL_ADC_PollForConversion(&hadc1, 1000);
-    float v_raw =  (HAL_ADC_GetValue(&hadc1) *3.3) / 4095.0; 
+    double v_raw =  (HAL_ADC_GetValue(&hadc1) *3.3) / 4095.0; 
 
     return v_raw * 1000;
 }
